Light default.py. New setting format
